+++ R/report.R
@@ -9,9 +9,8 @@
 #' report(x)
 #' }
 #' @export
-# This function was originally a shiny application, but has been converted into
-# a normal static document. Hence the shiny calls / dependency despite not
-# actually using shiny.
+# This function was originally a shiny application, but has now been converted into
+# a normal static document and no longer depends on shiny.
 report <- function(x = package_coverage(),
   file = file.path(tempdir(), paste0(get_package_name(x), "-report.html")),
   browse = interactive()) {
@@ -22,15 +21,14 @@ report <- function(x = package_coverage(),
   # Paths need to be absolute for save_html to work properly
   file <- file.path(normalizePath(dirname(file), mustWork = TRUE), basename(file))
 
-  loadNamespace("shiny")
+  loadNamespace("htmltools")
   loadNamespace("DT")
 
-  data <- to_shiny_data(x)
+  data <- to_report_data(x)
 
   # Color the td cells by coverage amount, like codecov.io does
   color_coverage_callback <- DT::JS(
 'function(td, cellData, rowData, row, col) {
-console.log(cellData);
   var percent = cellData.replace("%", "");
   if (percent > 90) {
     var grad = "linear-gradient(90deg, #edfde7 " + cellData + ", white " + cellData + ")";
@@ -56,27 +54,32 @@ console.log(cellData);
   package_name <- attr(x, "package")$package
   percentage <- sprintf("%02.2f%%", data$overall)
 
-  ui <- shiny::fluidPage(
-    shiny::includeCSS(system.file("www/shiny.css", package = "covr")),
+  table <- DT::datatable(
+    data$file_stats,
+    escape = FALSE,
+    fillContainer = TRUE,
+    options = list(
+      searching = FALSE,
+      dom = "t",
+      paging = FALSE,
+      columnDefs = list(
+        list(targets = 6, createdCell = color_coverage_callback))),
+    rownames = FALSE,
+    class = "row-border",
+    callback = file_choice_callback
+  )
+  table$sizingPolicy$defaultWidth <- "100%"
+  table$sizingPolicy$defaultHeight <- NULL
+
+  ui <- fluid_page(
+    htmltools::includeCSS(system.file("www/report.css", package = "covr")),
     column(8, offset = 2, size = "md",
-      shiny::HTML(paste0("<h2>", package_name, " coverage - ", percentage, "</h2>")),
-      shiny::tabsetPanel(
-        shiny::tabPanel("Files",
-          DT::datatable(
-            data$file_stats,
-            escape = FALSE,
-            options = list(
-              searching = FALSE,
-              dom = "t",
-              paging = FALSE,
-              columnDefs = list(
-                list(targets = 6, createdCell = color_coverage_callback))),
-            rownames = FALSE,
-            class = "row-border",
-            callback = file_choice_callback
-          )
+      htmltools::HTML(paste0("<h2>", package_name, " coverage - ", percentage, "</h2>")),
+      tabset_panel(
+        tab_panel("Files",
+          table
         ),
-        shiny::tabPanel("Source", addHighlight(renderSourceTable(data$full)))
+        tab_panel("Source", addHighlight(renderSourceTable(data$full)))
       )
     )
   )
@@ -97,7 +100,7 @@ console.log(cellData);
 #' @param out_file The output file
 #' @export
 file_report <- function(x = package_coverage(), file = NULL, out_file = file.path(tempdir(), paste0(get_package_name(x), "-file-report.html")), browse = interactive()) {
-  loadNamespace("shiny")
+  loadNamespace("htmltools")
   loadNamespace("DT")
 
   files <- display_name(x)
@@ -109,14 +112,14 @@ file_report <- function(x = package_coverage(), file = NULL, out_file = file.pat
 
   x <- x[files %in% file]
 
-  data <- to_shiny_data(x)
+  data <- to_report_data(x)
 
   percentage <- data$file_stats$Coverage
 
-  ui <- shiny::fluidPage(
-    shiny::includeCSS(system.file("www/shiny.css", package = "covr")),
+  ui <- fluid_page(
+    htmltools::includeCSS(system.file("www/report.css", package = "covr")),
     column(8, offset = 2, size = "md",
-      shiny::HTML(paste0("<h2>", file, " - ", percentage, "</h2>")),
+      htmltools::HTML(paste0("<h2>", file, " - ", percentage, "</h2>")),
       addHighlight(
         renderSourceTable(data$full, "")
       )
@@ -132,7 +135,7 @@ file_report <- function(x = package_coverage(), file = NULL, out_file = file.pat
   invisible(out_file)
 }
 
-to_shiny_data <- function(x) {
+to_report_data <- function(x) {
   coverages <- per_line(x)
 
   res <- list()
@@ -190,16 +193,16 @@ sort_file_stats <- function(stats) {
 }
 
 add_link <- function(files) {
-  vcapply(files, function(file) { as.character(shiny::a(href = "#", file)) })
+  vcapply(files, function(file) { as.character(htmltools::a(href = "#", file)) })
 }
 
 renderSourceTable <- function(data, class = "hidden") {
 
-  shiny::tags$div(id = "files",
+  htmltools::div(id = "files",
     Map(function(lines, file) {
-      shiny::tags$div(id = file, class=class,
-        shiny::tags$table(class = "table-condensed",
-          shiny::tags$tbody(
+      htmltools::div(id = file, class=class,
+        htmltools::tags$table(class = "table-condensed",
+          htmltools::tags$tbody(
             lapply(seq_len(NROW(lines)),
               function(row_num) {
                 coverage <- lines[row_num, "coverage"]
@@ -209,22 +212,22 @@ renderSourceTable <- function(data, class = "hidden") {
                   cov_value <- "!"
                   cov_type <- "missed"
                 } else if (coverage > 0) {
-                  cov_value <- shiny::HTML(paste0(lines[row_num, "coverage"], "<em>x</em>", collapse = ""))
+                  cov_value <- htmltools::HTML(paste0(lines[row_num, "coverage"], "<em>x</em>", collapse = ""))
                   cov_type <- "covered"
                 } else {
                   cov_type <- "never"
                   cov_value <- ""
                 }
-                shiny::tags$tr(class = cov_type,
-                  shiny::tags$td(class = "num", lines[row_num, "line"]),
-                  shiny::tags$td(class = "coverage", cov_value),
-                  shiny::tags$td(class = "col-sm-12", shiny::pre(class = "language-r", lines[row_num, "source"]))
+                htmltools::tags$tr(class = cov_type,
+                  htmltools::tags$td(class = "num", lines[row_num, "line"]),
+                  htmltools::tags$td(class = "coverage", cov_value),
+                  htmltools::tags$td(class = "col-sm-12", htmltools::pre(class = "language-r", lines[row_num, "source"]))
                   )
               })
             )
           ))
     }, lines = data, file = names(data)),
-  shiny::tags$script(
+  htmltools::tags$script(
     "$('div#files pre').each(function(i, block) {
     hljs.highlightBlock(block);
 });"))
@@ -232,8 +235,8 @@ renderSourceTable <- function(data, class = "hidden") {
 
 addHighlight <- function(x = list()) {
   highlight <- htmltools::htmlDependency("highlight.js", "6.2",
-                                         system.file(package = "shiny",
-                                                     "www/shared/highlight"),
+                                         system.file(package = "covr",
+                                                     "www/shared/highlight.js"),
                                          script = "highlight.pack.js",
                                          stylesheet = "rstudio.css")
 
@@ -248,6 +251,8 @@ addin_report <- function() {
   covr::report(covr::package_coverage(project %||% getwd()))
 }
 
+# These are all adapted from functions in shiny
+
 column <- function(width, ..., offset = 0, size = c("xs", "sm", "md", "lg")) {
   size <- match.arg(size)
 
@@ -255,5 +260,67 @@ column <- function(width, ..., offset = 0, size = c("xs", "sm", "md", "lg")) {
   if (offset > 0) {
     col_class <- paste0(col_class, " ", "col-", size, "-offset-", offset)
   }
-  shiny::div(class = col_class, ...)
+  htmltools::div(class = col_class, ...)
 }
+
+tab_panel <- function(title, ..., value = title) {
+  htmltools::div(class = "tab-pane", title = title, `data-value` = value, ...)
+}
+
+fluid_page <- function(...) {
+  bootstrap_page(
+    htmltools::div(class = "container-fluid", ...)
+  )
+}
+
+bootstrap_page <- function(...) {
+  htmltools::attachDependencies(htmltools::tagList(list(...)), html_dependency_bootstrap())
+}
+
+# from htmldeps::html_dependency_bootstrap (not yet on CRAN)
+html_dependency_bootstrap <- function () {
+  htmltools::htmlDependency(name = "bootstrap", version = "3.3.5",
+    src = system.file(file = "www/shared/bootstrap", package = "covr"),
+    meta = list(viewport = "width=device-width, initial-scale=1"),
+    script = c("js/bootstrap.min.js", "shim/html5shiv.min.js", "shim/respond.min.js"),
+    stylesheet = c("css/bootstrap.min.css", "css/bootstrap-theme.min.css")
+  )
+}
+
+tabset_panel <- function(...) {
+  tabset <- build_tabset(list(...))
+  htmltools::div(class = "tabbable",
+    tabset$nav_list,
+    tabset$content)
+}
+
+build_tabset <- function(tabs) {
+  tabset_id <- "covr"
+  tabs <- lapply(seq_len(length(tabs)), build_tab_item, tabs = tabs, tabset_id = tabset_id)
+  list(nav_list = ul(class = "nav nav-tabs", `data-tabsetid` = tabset_id, lapply(tabs, "[[", 1)),
+       content = htmltools::div(class = "tab-content", `data-tabsetid` = tabset_id, lapply(tabs, "[[", 2))
+  )
+}
+
+build_tab_item <- function(i, tabs, tabset_id) {
+
+  div_tag <- tabs[[i]]
+  tab_id <- paste("tab", tabset_id, i, sep = "-")
+  li_tag <- li(
+    htmltools::a(href = paste0("#", tab_id),
+      `data-toggle` = "tab",
+      `data-value` = div_tag$attribs$`data-value`,
+      div_tag$attribs$title
+    )
+  )
+  if (i == 1) {
+    li_tag$attribs$class <- "active"
+    div_tag$attribs$class <- paste(div_tag$attribs$class, "active")
+  }
+  div_tag$attribs$id <- tab_id
+
+  list(li_tag = li_tag, div_tag = div_tag)
+}
+
+li <- function(...) htmltools::tag("li", list(...))
+ul <- function(...) htmltools::tag("ul", list(...))